+++ lib/Epub/Epub/ReaderRenderSpec.h
@@ -23,7 +23,7 @@ struct ReaderRenderSpec {
   // 0 disables hanging punctuation; otherwise maximum overhang into the physical margin.
   uint8_t hangingPunctuationLimitPx = 0;
   bool fixedDialogueSpacing = false;
-  uint8_t letterSpacingLimitPercent = 0;
+  uint16_t letterSpacingLimitPercent = 0;
   uint8_t minimumSpacePercent = 100;
   bool embeddedStyle = true;
   uint8_t imageRendering = 0;
